+++ packages/api/src/__package__.ts
@@ -1,3 +1,3 @@
 export const PKG_NAME = "@hpcc-js/api";
-export const PKG_VERSION = "2.7.7";
-export const BUILD_VERSION = "2.6.7";
+export const PKG_VERSION = "2.7.8";
+export const BUILD_VERSION = "2.6.8";