+++ packages/comms/src/ecl/targetCluster.ts
@@ -1,13 +1,13 @@
 import { Cache, StateObject } from "@hpcc-js/util";
 import { IConnection, IOptions } from "../connection";
 import { GetTargetClusterInfo, GetTargetClusterUsageEx, MachineService } from "../services/wsMachine";
-import { TopologyService, TpTargetClusterQuery } from "../services/wsTopology";
+import { TopologyService, TpListTargetClusters, TpTargetClusterQuery } from "../services/wsTopology";
 import { Machine } from "./machine";
 
-export class TargetClusterCache extends Cache<{ Name: string }, TargetCluster> {
+export class TargetClusterCache extends Cache<{ BaseUrl: string, Name: string }, TargetCluster> {
     constructor() {
         super((obj) => {
-            return obj.Name;
+            return `${obj.BaseUrl}-${obj.Name}`;
         });
     }
 }
@@ -17,15 +17,17 @@ export interface TpTargetClusterEx {
     MachineInfoEx: GetTargetClusterInfo.MachineInfoEx[];
 }
 
-export type UTargetClusterState = TpTargetClusterQuery.TpTargetCluster & TpTargetClusterEx;
-export type ITargetClusterState = TpTargetClusterQuery.TpTargetCluster | TpTargetClusterEx;
+export type UTargetClusterState = TpTargetClusterQuery.TpTargetCluster & TpListTargetClusters.TpClusterNameType & TpTargetClusterEx;
+export type ITargetClusterState = TpTargetClusterQuery.TpTargetCluster | TpListTargetClusters.TpClusterNameType | TpTargetClusterEx;
 export class TargetCluster extends StateObject<UTargetClusterState, ITargetClusterState> implements UTargetClusterState {
     protected connection: TopologyService;
     protected machineConnection: MachineService;
+    get BaseUrl() { return this.connection.baseUrl; }
 
     get Name(): string { return this.get("Name"); }
     get Prefix(): string { return this.get("Prefix"); }
     get Type(): string { return this.get("Type"); }
+    get IsDefault(): boolean { return this.get("IsDefault"); }
     get TpClusters(): TpTargetClusterQuery.TpClusters { return this.get("TpClusters"); }
     get TpEclCCServers(): TpTargetClusterQuery.TpEclCCServers { return this.get("TpEclCCServers"); }
     get TpEclServers(): TpTargetClusterQuery.TpEclServers { return this.get("TpEclServers"); }
@@ -37,7 +39,7 @@ export class TargetCluster extends StateObject<UTargetClusterState, ITargetClust
     }
 
     static attach(optsConnection: IOptions | IConnection | TopologyService, name: string, state?: ITargetClusterState): TargetCluster {
-        const retVal: TargetCluster = _targetCluster.get({ Name: name }, () => {
+        const retVal: TargetCluster = _targetCluster.get({ BaseUrl: optsConnection.baseUrl, Name: name }, () => {
             return new TargetCluster(optsConnection, name);
         });
         if (state) {
@@ -103,4 +105,47 @@ export class TargetCluster extends StateObject<UTargetClusterState, ITargetClust
     fetchUsage(): Promise<GetTargetClusterUsageEx.TargetClusterUsage[]> {
         return this.machineConnection.GetTargetClusterUsageEx([this.Name]);
     }
+}
+
+export function targetClusters(optsConnection: IOptions | IConnection | TopologyService): Promise<TargetCluster[]> {
+    let connection: TopologyService;
+    if (optsConnection instanceof TopologyService) {
+        connection = optsConnection;
+    } else {
+        connection = new TopologyService(optsConnection);
+    }
+    return connection.TpListTargetClusters({}).then(response => {
+        return response.TargetClusters.TpClusterNameType.map(item => TargetCluster.attach(optsConnection, item.Name, item));
+    });
+}
+
+const _defaultTargetCluster: { [baseUrl: string]: Promise<TargetCluster> } = {};
+export function defaultTargetCluster(optsConnection: IOptions | IConnection | TopologyService): Promise<TargetCluster> {
+    if (!_defaultTargetCluster[optsConnection.baseUrl]) {
+        let connection: TopologyService;
+        if (optsConnection instanceof TopologyService) {
+            connection = optsConnection;
+        } else {
+            connection = new TopologyService(optsConnection);
+        }
+        _defaultTargetCluster[optsConnection.baseUrl] = connection.TpListTargetClusters({}).then(response => {
+            let firstItem: TpListTargetClusters.TpClusterNameType;
+            let defaultItem: TpListTargetClusters.TpClusterNameType;
+            let hthorItem: TpListTargetClusters.TpClusterNameType;
+            response.TargetClusters.TpClusterNameType.forEach(item => {
+                if (!firstItem) {
+                    firstItem = item;
+                }
+                if (!defaultItem && item.IsDefault === true) {
+                    defaultItem = item;
+                }
+                if (!hthorItem && item.Type === "hthor") {
+                    hthorItem = item;
+                }
+            });
+            const defItem = defaultItem || hthorItem || firstItem;
+            return TargetCluster.attach(optsConnection, defItem.Name, defItem);
+        });
+    }
+    return _defaultTargetCluster[optsConnection.baseUrl];
 }