+++ packages/eclwatch/src/__package__.ts
@@ -1,3 +1,3 @@
 export const PKG_NAME = "@hpcc-js/eclwatch";
-export const PKG_VERSION = "2.5.7";
-export const BUILD_VERSION = "2.6.7";
+export const PKG_VERSION = "2.5.8";
+export const BUILD_VERSION = "2.6.8";