+++ demos/gallery/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.23.2](https://github.com/GordonSmith/Visualization/compare/@hpcc-js/demo-gallery@2.23.1...@hpcc-js/demo-gallery@2.23.2) (2019-05-16)
+
+**Note:** Version bump only for package @hpcc-js/demo-gallery
+
+
+
+
+
+
 ## [2.23.1](https://github.com/GordonSmith/Visualization/compare/@hpcc-js/demo-gallery@2.23.0...@hpcc-js/demo-gallery@2.23.1) (2019-04-27)
 
 **Note:** Version bump only for package @hpcc-js/demo-gallery