+++ tests/test-composite/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.33](https://github.com/GordonSmith/Visualization/compare/@hpcc-js/test-composite@2.0.32...@hpcc-js/test-composite@2.0.33) (2019-05-16)
+
+**Note:** Version bump only for package @hpcc-js/test-composite
+
+
+
+
+
+
 ## [2.0.32](https://github.com/GordonSmith/Visualization/compare/@hpcc-js/test-composite@2.0.31...@hpcc-js/test-composite@2.0.32) (2019-04-27)
 
 **Note:** Version bump only for package @hpcc-js/test-composite