+++ tests/test-graph/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.30](https://github.com/GordonSmith/Visualization/compare/@hpcc-js/test-graph@2.0.29...@hpcc-js/test-graph@2.0.30) (2019-05-16)
+
+**Note:** Version bump only for package @hpcc-js/test-graph
+
+
+
+
+
+
 ## [2.0.29](https://github.com/GordonSmith/Visualization/compare/@hpcc-js/test-graph@2.0.28...@hpcc-js/test-graph@2.0.29) (2019-04-27)
 
 **Note:** Version bump only for package @hpcc-js/test-graph