+++ tests/test-other/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.26](https://github.com/GordonSmith/Visualization/compare/@hpcc-js/test-other@2.1.25...@hpcc-js/test-other@2.1.26) (2019-05-16)
+
+**Note:** Version bump only for package @hpcc-js/test-other
+
+
+
+
+
+
 ## [2.1.25](https://github.com/GordonSmith/Visualization/compare/@hpcc-js/test-other@2.1.24...@hpcc-js/test-other@2.1.25) (2019-04-27)
 
 **Note:** Version bump only for package @hpcc-js/test-other